@@ -1,9 +1,11 @@
 module github.com/mmcloughlin/avo
 
-go 1.16
+go 1.18
 
 require (
 	golang.org/x/arch v0.2.0
-	golang.org/x/sys v0.3.0
-	golang.org/x/tools v0.4.0
+	golang.org/x/sys v0.5.0
+	golang.org/x/tools v0.6.0
 )
+
+require golang.org/x/mod v0.8.0 // indirect
