@@ -1,5 +1,4 @@
 //go:build ignore
-// +build ignore
 
 package main
 
@@ -102,7 +101,7 @@ func GenerateWorkflow(s *thirdparty.Suite) ([]byte, error) {
 		g.Linef("- name: Install Go")
 		g.Linef("  uses: actions/setup-go@c4a742cab115ed795e34d4513e2cf7d472deb55f # v3.3.1")
 		g.Linef("  with:")
-		g.Linef("    go-version: 1.19.x")
+		g.Linef("    go-version: 1.20.x")
 		g.Linef("    check-latest: true")
 
 		// Checkout avo.
